requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
Aspose.Slides==24.7.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
dataset==1.6.2
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
efficientnet-pytorch==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imgaug==0.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
lightning==2.3.3
lightning-utilities==0.11.3.post0
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@5360b60fad91cc54b418820c2d09f188cfd6abd0
lmdb==1.5.1
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.35.14
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.0
paddlepaddle==2.6.1
pandas==2.2.2
parso==0.8.4
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.2
peft==0.11.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytorch-lightning==2.3.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.52
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.0
tbb==2021.11.0
termcolor==2.4.0
tifffile==2024.7.2
tokenizers==0.19.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
widgetsnbextension==4.0.11
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
